Rename blockexplorer module to explorer
